requirements.txt
fastapi
uvicorn
numpy
opencv-python
pytesseract
textdistance
pyspellchecker
python-multipart
opencv-contrib-python 
caer
pillow
google-cloud-vision
google-api-core
google-auth

